Ignore non-directories when running rfs query
Update sweeps_utils.py collect_rf_statis() to check whether a path is a directory before running check_status().

This has been added due to errors arising when macOS adds a hidden ".DS_Store" file to ./rfs/, breaking the query function.

Now non-directories in ./rfs/ are ignored. If a file is not hidden, a warning statement is printed.