Correct sweeps_analysis JLD get_data multiple keys
The get_data() functionality for HDF5, JLD, and JLD2 filetypes has been
corrected to send only the "filename" path to the  HDF5 file. The h5py
Python package does not support sending a h5py objet directly, so the
filename is sent to the caller for it to read the data using the h5py
interface, for example:

import h5py
with h5py.File(get_data(ID, sim_loc), 'r') as f:
    object = f.get('key_name')[()])

An exception to this usage is if the filetype is a JLD or JLD2 file with
only one key, taken to represent an array which is converted to a NumPy
array and returned as such.